Рефакторинг структуры компонентов
