+++ src/components/common/rating-input/rating-input.tsx
