--- src/components/favorites-list/favorites-list.tsx
@@ -1,11 +1,0 @@
-import FavoritesItem from '../favorites-item/favorites-item';
-
-function FavoritesList(): JSX.Element {
-  return (
-    <ul className="favorites__list">
-      <FavoritesItem />
-    </ul>
-  );
-}
-
-export default FavoritesList;
