+++ src/components/favorites-screen/favorites-empty/favorites-empty.tsx
