+++ src/components/favorites-screen/favorites-list/favorites-list.tsx
@@ -1,0 +1,11 @@
+import FavoritesItem from '../favorites-item/favorites-item.tsx';
+
+function FavoritesList(): JSX.Element {
+  return (
+    <ul className="favorites__list">
+      <FavoritesItem />
+    </ul>
+  );
+}
+
+export default FavoritesList;
