--- src/components/offer-components/nearby-offers/nearby-offers.tsx
@@ -1,30 +1,0 @@
-import { TOffers } from '../../../types/offer';
-import OffersList from '../../offers-list/offers-list';
-
-type NearbyOffersProps = {
-  slicedNearbyOffers: TOffers;
-};
-
-function NearbyOffers({ slicedNearbyOffers }: NearbyOffersProps): JSX.Element {
-  if (slicedNearbyOffers.length === 0) {
-    return (
-      <section className="near-places places">
-        <h2 className="near-places__title">
-          Other places in the neighborhood not found, sorry
-        </h2>
-      </section>
-    );
-  }
-
-  return (
-    <section className="near-places places">
-      <h2 className="near-places__title">Other places in the neighborhood</h2>
-      <OffersList
-        offers={slicedNearbyOffers}
-        className="near-places__list places__list"
-      />
-    </section>
-  );
-}
-
-export default NearbyOffers;
