--- src/components/offer-components/offer-price/offer-price.tsx
@@ -1,16 +1,0 @@
-import { TOffer } from '../../../types/offer.ts';
-
-type OfferPriceProps = {
-  price: TOffer['price'];
-};
-
-function OfferPrice({ price }: OfferPriceProps): JSX.Element {
-  return (
-    <div className="offer__price">
-      <b className="offer__price-value">&euro;{price}</b>
-      <span className="offer__price-text">&nbsp;night</span>
-    </div>
-  );
-}
-
-export default OfferPrice;
