+++ src/components/offer-screen/offer/offer-features/offer-features.tsx
