+++ src/components/offer-screen/offer/offer-inside/offer-inside.tsx
