+++ src/index.tsx
@@ -5,7 +5,7 @@ import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import browserHistory from './browser-history';
 import App from './components/app';
-import HistoryRouter from './components/history-route/history-route';
+import HistoryRouter from './components/routes/history-route/history-route';
 import { store } from './store';
 import { fetchOffersAction, checkAuthAction } from './store/api-actions';
