+++ src/pages/favorites-screen/favorites-screen.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { Helmet } from 'react-helmet-async';
-import FavoritesEmpty from '../../components/favorites-empty/favorites-empty';
-import FavoritesNotEmpty from '../../components/favorites-not-empty/favorites-not-empty';
+import FavoritesEmpty from '../../components/favorites-screen/favorites-empty/favorites-empty';
+import FavoritesNotEmpty from '../../components/favorites-screen/favorites-not-empty/favorites-not-empty';
 import Footer from '../../components/ui/footer/footer';
 import Header from '../../components/ui/header/header';
 import { useAppDispatch, useAppSelector } from '../../hooks';
